+++ test/4.x/app.js
@@ -235,7 +235,7 @@ test('helper error', function (done) {
 test('partials', function(done) {
   request(app)
     .get('/partials')
-    .expect(shouldHaveFirstLineEqual('Test Partial 1Test Partial 2Test Partial 3Test Partial 4'))
+    .expect(shouldHaveFirstLineEqual('Test Partial 1Test Partial 2Test Partial 3Test Partial 4Test Partial 5'))
     .end(done)
 });